Abstract tests, add insert ID support, add MySQLI driver.
Generic SQL driver tests are now in SQLExecutor, a superclass to
MySQLExecutor and MySQLIExecutor, so that they do not need to be
duplicated.

Add #getLastInsertId() method to SQLResult.

Change argument order to BasicSQLResult so that rows are first,
followed by affected row count and then insert ID.

Both SQLExecutor tests call <tested class>::createFromConfig directly
instead of using the driver named in the config (as relying on that
sort of defeats the purpose).

Possibly other minor refactoring/clean-up.